Split types.ts into multiple files
