+++ packages/fastify-vite/src/types/client.ts
@@ -1,0 +1,11 @@
+export interface ClientModule {
+  createApp?: (...args: unknown[]) => unknown
+  create?: (...args: unknown[]) => unknown
+  routes?: unknown
+  [key: string]: unknown
+}
+
+export interface ClientEntries {
+  ssr?: ClientModule
+  [key: string]: unknown
+}
