+++ packages/fastify-vite/src/types/reply.ts
@@ -1,0 +1,34 @@
+import type { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify'
+
+import type { RouteDefinition } from './route.ts'
+
+export interface ReplyDotRenderContext {
+  app?: FastifyInstance
+  server?: FastifyInstance
+  req?: FastifyRequest
+  reply?: FastifyReply
+  client?: unknown
+  route?: RouteDefinition
+  [key: string]: unknown
+}
+
+export type ReplyDotRenderResult = Record<string, unknown>
+
+export type ReplyDotRenderFunction = (
+  this: FastifyReply,
+  ctx?: ReplyDotRenderContext,
+) => ReplyDotRenderResult | Promise<ReplyDotRenderResult>
+
+export type ReplyDotHtmlFunction = (
+  this: FastifyReply,
+  ctx?: ReplyDotRenderResult,
+) => FastifyReply | Promise<FastifyReply>
+
+/**
+ * A FastifyReply that has been decorated with html() and render() methods.
+ * This is the reply type available after vite.ready() has been called.
+ */
+export interface DecoratedReply extends FastifyReply {
+  render: (ctx?: ReplyDotRenderContext) => ReplyDotRenderResult | Promise<ReplyDotRenderResult>
+  html: (ctx?: ReplyDotRenderResult) => FastifyReply | Promise<FastifyReply>
+}
